Reorganized dirs.
